fix: Namespace von FriendsOfREDAXO auf FriendsOfRedaxo korrigiert

+++ pages/system.activity-log.php
@@ -3,8 +3,9 @@
 /** @var \rex_addon $this */
 
 $table = \rex::getTable('activity_log');
+$isAdmin = \rex::getUser()->isAdmin();
 
-if (rex_post('delete_old_logs') && 1 == rex_post('delete_old_logs')) {
+if ($isAdmin && rex_post('delete_old_logs') && 1 == rex_post('delete_old_logs')) {
     $now = (new \DateTime());
     $now->modify('-7 day');
     $date = $now->format('Y-m-d H:i:s');
@@ -15,13 +16,13 @@ if (rex_post('delete_old_logs') && 1 == rex_post('delete_old_logs')) {
     $sql->delete();
 }
 
-if (rex_post('delete_all_logs') && 1 == rex_post('delete_all_logs')) {
+if ($isAdmin && rex_post('delete_all_logs') && 1 == rex_post('delete_all_logs')) {
     $sql = \rex_sql::factory();
     $sql->setTable($table);
     $sql->delete();
 }
 
-if (rex_post('delete_single_log')) {
+if ($isAdmin && rex_post('delete_single_log')) {
     $sql = \rex_sql::factory();
     $sql->setTable($table);
     $sql->setWhere('id = ' . rex_post('delete_single_log'));
@@ -83,10 +84,12 @@ $list->setColumnFormat('message', 'custom', 'rex_activity::messageListCallback')
 $list->setColumnFormat('type', 'custom', 'rex_activity::typeListCallback');
 $list->setColumnFormat('source', 'custom', 'rex_activity::sourceListCallback');
 
-$list->addColumn('delete', '', -1, ['<th></th>', '<td class="rex-table-icon">###VALUE###</td>']);
-$list->setColumnFormat('delete', 'custom', static function ($params) use ($list) {
-    return '<button type="submit" class="btn btn-danger btn-sm" name="delete_single_log" value="' . $list->getValue('id') . '"><i class="rex-icon rex-icon-delete"></i></button>';
-});
+if ($isAdmin) {
+    $list->addColumn('delete', '', -1, ['<th></th>', '<td class="rex-table-icon">###VALUE###</td>']);
+    $list->setColumnFormat('delete', 'custom', static function ($params) use ($list) {
+        return '<button type="submit" class="btn btn-danger btn-sm" name="delete_single_log" value="' . $list->getValue('id') . '"><i class="rex-icon rex-icon-delete"></i></button>';
+    });
+}
 
 // Filter-Parameter durch Pager-Links und List-Form (delete_single_log) durchreichen
 if ('' !== $type)   { $list->addParam('type',   $type); }
@@ -111,10 +114,12 @@ foreach (['type' => $type, 'user' => $user, 'source' => $source, 'search' => $se
     }
 }
 
-$content .= '<hr><form method="post" action="' . \rex_url::currentBackendPage() . '" style="text-align: right">' . $filterHidden . '
-    <button type="submit" class="btn btn-danger" name="delete_old_logs" value="1">' . $addon->i18n('delete_older_than_7_days') . '</button>
-    <button type="submit" class="btn btn-danger" name="delete_all_logs" value="1">' . $addon->i18n('delete_all') . '</button>
-</form>';
+if ($isAdmin) {
+    $content .= '<hr><form method="post" action="' . \rex_url::currentBackendPage() . '" style="text-align: right">' . $filterHidden . '
+        <button type="submit" class="btn btn-danger" name="delete_old_logs" value="1">' . $addon->i18n('delete_older_than_7_days') . '</button>
+        <button type="submit" class="btn btn-danger" name="delete_all_logs" value="1">' . $addon->i18n('delete_all') . '</button>
+    </form>';
+}
 
 $fragment = new \rex_fragment();
 $fragment->setVar('body', $content, false);
